Fix Dhrystone metric overflow, EMBench wikisort build, and path portability

+++ dhrystone/dhrystone_support.c
@@ -13,18 +13,25 @@ void* malloc(size_t size) {
     return NULL;
 }
 
-/* 2. Timer functions using RISC-V CSRs */
+/*
+ * 2. Timer functions using RISC-V mcycle / minstret CSRs.
+ * We read only the 32-bit low halves — Dhrystone with 100,000 iterations
+ * takes ~144M cycles, which fits in 32 bits (max ~4.29B).
+ * The metric calculation in dhry_1.c uses explicit (long long) casts
+ * to avoid overflow in the intermediate products.
+ */
 long time(long* tloc) {
     uint32_t cycles;
     __asm__ volatile ("csrr %0, mcycle" : "=r" (cycles));
-    if (tloc) *tloc = cycles;
-    return cycles;
+    if (tloc) *tloc = (long)cycles;
+    return (long)cycles;
 }
 
 long insn(long* tloc) {
     uint32_t instret;
     __asm__ volatile ("csrr %0, minstret" : "=r" (instret));
-    if (tloc) *tloc = instret;
-    return instret;
+    if (tloc) *tloc = (long)instret;
+    return (long)instret;
 }
+
